Restructure KAN and NBeats layers to include them in pytorch_forecasting/layers for better maintainability.
