+++ pytorch_forecasting/layers/_nbeats/__init__.py
@@ -1,0 +1,17 @@
+"""
+Implementation of N-BEATS model blocks and utilities.
+"""
+
+from pytorch_forecasting.layers._nbeats._blocks import (
+    NBEATSBlock,
+    NBEATSGenericBlock,
+    NBEATSSeasonalBlock,
+    NBEATSTrendBlock,
+)
+
+__all__ = [
+    "NBEATSBlock",
+    "NBEATSGenericBlock",
+    "NBEATSSeasonalBlock",
+    "NBEATSTrendBlock",
+]
