+++ pytorch_forecasting/models/nbeats/__init__.py
@@ -6,20 +6,12 @@ from pytorch_forecasting.models.nbeats._nbeats_adapter import NBeatsAdapter
 from pytorch_forecasting.models.nbeats._nbeats_pkg import NBeats_pkg
 from pytorch_forecasting.models.nbeats._nbeatskan import NBeatsKAN
 from pytorch_forecasting.models.nbeats._nbeatskan_pkg import NBeatsKAN_pkg
-from pytorch_forecasting.models.nbeats.sub_modules import (
-    NBEATSGenericBlock,
-    NBEATSSeasonalBlock,
-    NBEATSTrendBlock,
-)
 
 __all__ = [
     "NBeats",
     "NBeatsKAN",
-    "NBEATSGenericBlock",
     "NBeats_pkg",
     "NBeatsKAN_pkg",
-    "NBEATSSeasonalBlock",
-    "NBEATSTrendBlock",
     "NBeatsAdapter",
     "GridUpdateCallback",
 ]
